refactor: socket-only NCCL check, add ProxyOp metrics, rename metric
- Remove file-based parser (parser.go, parser_stub.go, parser_test.go)
- Move types to types.go (NCCLInspectorEvent, ParsedEvent, parseEvent)
- Socket-only collection (no file fallback)
- Add ProxyOp network transfer metrics (nccl.proxy_op.max_network_time_us)
- Rename metric from network.max_transfer_time_us to proxy_op.max_network_time_us
- Add debug logging for coll_perf and proxy_op events
- Refactor mutator to use mutatecommon helpers, support two .so files
