--- pkg/collector/corechecks/gpu/nccl/parser_stub.go
@@ -1,11 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2024-present Datadog, Inc.
-
-//go:build !linux
-
-package nccl
-
-// Stub file for non-Linux platforms
-// Parser types are only needed on Linux where the check runs
